:ambulance: Fix initialization in bit_bang_i2c constructor
Critical bug fix in bit_bang_i2c that can lead to device damage. The
output pins NEED to be set as open drain in order to not cause bus
contention which can damage devices and cause short circuits.

The constructor now properly configures the sda and scl output pins and
configures and performs initial configuration.

Resolves #36